+++ runtime/tests/integration_tests.rs
@@ -3,7 +3,7 @@
 use caliptra_builder::{FwId, ImageOptions, APP_WITH_UART, FMC_WITH_UART, ROM_WITH_UART};
 use caliptra_drivers::Ecc384PubKey;
 use caliptra_hw_model::{BootParams, DefaultHwModel, HwModel, InitParams, ModelError, ShaAccMode};
-use caliptra_runtime::{CommandId, EcdsaVerifyCmd, FipsModuleApi};
+use caliptra_runtime::{CommandId, EcdsaVerifyCmd};
 use openssl::{
     bn::BigNum,
     ec::{EcGroup, EcKey},
@@ -226,18 +226,18 @@ fn test_fips_cmd_api() {
 
     let cmd = [0u8; 4];
 
-    let resp = model.mailbox_execute(u32::from(FipsModuleApi::VERSION), &cmd);
+    let resp = model.mailbox_execute(u32::from(CommandId::VERSION), &cmd);
     assert_eq!(resp, expected_err);
 
-    let resp = model.mailbox_execute(u32::from(FipsModuleApi::SHUTDOWN), &cmd);
+    let resp = model.mailbox_execute(u32::from(CommandId::SHUTDOWN), &cmd);
     assert_eq!(resp, expected_err);
 
-    let resp = model.mailbox_execute(u32::from(FipsModuleApi::SELF_TEST), &cmd);
+    let resp = model.mailbox_execute(u32::from(CommandId::SELF_TEST), &cmd);
     assert_eq!(resp, expected_err);
 
     let expected_err = Err(ModelError::MailboxCmdFailed(0xe0002));
     // Send something that is not a valid RT command.
-    let resp = model.mailbox_execute(0xBABEFACE, &cmd);
+    let resp = model.mailbox_execute(0xAABBCCDD, &cmd);
     assert_eq!(resp, expected_err);
 }
